Setup logging earlier
