+++ server/config.go
@@ -20,6 +20,8 @@ import (
 	"fmt"
 	"strconv"
 	"time"
+
+	"github.com/sirupsen/logrus"
 )
 
 // Environment represents an abstraction over the environment from which Service Manager configuration will be loaded
@@ -111,6 +113,8 @@ func NewConfiguration(env Environment) (*Config, error) {
 		config.LogLevel = configSettings.Log.Level
 	}
 
+	setUpLogging(config.LogLevel, config.LogFormat)
+
 	return config, nil
 }
 
@@ -135,4 +139,17 @@ func (c *Config) Validate() error {
 		return fmt.Errorf("validate Config: DbURI missing")
 	}
 	return nil
+}
+
+func setUpLogging(logLevel string, logFormat string) {
+	level, err := logrus.ParseLevel(logLevel)
+	if err != nil {
+		logrus.Fatal("Could not parse log level configuration")
+	}
+	logrus.SetLevel(level)
+	if logFormat == "json" {
+		logrus.SetFormatter(&logrus.JSONFormatter{})
+	} else {
+		logrus.SetFormatter(&logrus.TextFormatter{})
+	}
 }
